# backend/openapi.yaml
openapi: 3.0.3
info:
  title: Order Management Dashboard API
  version: 1.0.0
  description: API para autenticación, consulta de órdenes, filtros y estadísticas del dashboard de órdenes.

servers:
  - url: http://localhost:8080
    description: Local development

tags:
  - name: Health
    description: Health check
  - name: Auth
    description: Autenticación y sesión
  - name: Orders
    description: Consulta de órdenes
  - name: Stats
    description: Métricas agregadas

paths:
  /health:
    get:
      tags:
        - Health
      summary: Health check
      operationId: healthCheck
      responses:
        "200":
          description: API funcionando correctamente
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthResponse"
              example:
                status: ok

  /api/auth/login:
    post:
      tags:
        - Auth
      summary: Iniciar sesión
      operationId: login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
            example:
              email: admin@test.com
              password: admin123
      responses:
        "200":
          description: Login exitoso
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LoginResponse"
        "400":
          description: Request inválido
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Credenciales inválidas
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /api/auth/refresh:
    post:
      tags:
        - Auth
      summary: Renovar access token
      operationId: refreshToken
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RefreshRequest"
            example:
              refresh_token: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
      responses:
        "200":
          description: Access token renovado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RefreshResponse"
        "400":
          description: Request inválido
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Refresh token inválido o expirado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /api/auth/logout:
    post:
      tags:
        - Auth
      summary: Cerrar sesión
      operationId: logout
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LogoutRequest"
            example:
              refresh_token: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
      responses:
        "200":
          description: Logout exitoso
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
              example:
                message: logout successful
        "400":
          description: Request inválido
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /api/orders:
    get:
      tags:
        - Orders
      summary: Listar órdenes
      operationId: listOrders
      security:
        - bearerAuth: []
      parameters:
        - name: page
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            default: 1
          example: 1
        - name: pageSize
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
          example: 10
        - name: canal
          in: query
          required: false
          schema:
            type: string
          example: WEB
        - name: company
          in: query
          required: false
          schema:
            type: string
          example: LP
        - name: fulfillmentType
          in: query
          required: false
          schema:
            type: string
          example: Fulfillment_Type_Liverpool
        - name: productType
          in: query
          required: false
          schema:
            type: string
          example: Soft Line
      responses:
        "200":
          description: Lista paginada de órdenes
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ListOrdersResponse"
        "400":
          description: Parámetros inválidos
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: No autorizado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /api/orders/filters:
    get:
      tags:
        - Orders
      summary: Obtener opciones de filtros
      operationId: getOrderFilters
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Opciones disponibles para filtros
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OrderFiltersOptions"
        "401":
          description: No autorizado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /api/stats:
    get:
      tags:
        - Stats
      summary: Obtener métricas agregadas
      operationId: getStats
      security:
        - bearerAuth: []
      parameters:
        - name: canal
          in: query
          required: false
          schema:
            type: string
          example: WEB
        - name: company
          in: query
          required: false
          schema:
            type: string
          example: LP
        - name: fulfillmentType
          in: query
          required: false
          schema:
            type: string
          example: Fulfillment_Type_Liverpool
        - name: productType
          in: query
          required: false
          schema:
            type: string
          example: Soft Line
      responses:
        "200":
          description: Métricas agregadas
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StatsResponse"
        "401":
          description: No autorizado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  schemas:
    HealthResponse:
      type: object
      properties:
        status:
          type: string
          example: ok
      required:
        - status

    MessageResponse:
      type: object
      properties:
        message:
          type: string
          example: logout successful
      required:
        - message

    ErrorResponse:
      type: object
      properties:
        error:
          type: object
          properties:
            code:
              type: string
              example: INVALID_REQUEST
            message:
              type: string
              example: request is invalid
          required:
            - code
            - message
      required:
        - error

    User:
      type: object
      properties:
        id:
          type: string
          example: 7f5bca0f-6f33-4872-83e2-9fd08b948111
        email:
          type: string
          format: email
          example: admin@test.com
        name:
          type: string
          example: Admin
      required:
        - id
        - email
        - name

    LoginRequest:
      type: object
      properties:
        email:
          type: string
          format: email
          example: admin@test.com
        password:
          type: string
          format: password
          example: admin123
      required:
        - email
        - password

    LoginResponse:
      type: object
      properties:
        access_token:
          type: string
          example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
        refresh_token:
          type: string
          example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
        user:
          $ref: "#/components/schemas/User"
      required:
        - access_token
        - refresh_token
        - user

    RefreshRequest:
      type: object
      properties:
        refresh_token:
          type: string
          example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
      required:
        - refresh_token

    RefreshResponse:
      type: object
      properties:
        access_token:
          type: string
          example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
      required:
        - access_token

    LogoutRequest:
      type: object
      properties:
        refresh_token:
          type: string
          example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
      required:
        - refresh_token

    Order:
      type: object
      properties:
        id:
          type: string
          example: 00ESOmGuuX1VGqlFH4FX
        no_pedido:
          type: string
          example: "6050000169"
        canal:
          type: string
          example: WEB
        sku:
          type: string
          example: "1032003148"
        fecha_estimada:
          type: string
          example: Por ahora no es posible mostrar la fecha de entrega.
        fulfillment_type:
          type: string
          nullable: true
          example: Fulfillment_Type_Liverpool
        product_type:
          type: string
          nullable: true
          example: Soft Line
        cantidad:
          type: integer
          nullable: true
          example: 1
        fecha_compra:
          type: string
          format: date-time
          nullable: true
          example: "2024-07-30T15:29:18Z"
        company:
          type: string
          example: LP
        has_error:
          type: boolean
          example: false
      required:
        - id
        - no_pedido
        - canal
        - sku
        - fecha_estimada
        - fulfillment_type
        - product_type
        - cantidad
        - fecha_compra
        - company
        - has_error

    Pagination:
      type: object
      properties:
        page:
          type: integer
          example: 1
        page_size:
          type: integer
          example: 10
        total:
          type: integer
          format: int64
          example: 59848
        total_pages:
          type: integer
          example: 5985
      required:
        - page
        - page_size
        - total
        - total_pages

    ListOrdersResponse:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/Order"
        pagination:
          $ref: "#/components/schemas/Pagination"
      required:
        - data
        - pagination

    OrderFiltersOptions:
      type: object
      properties:
        channels:
          type: array
          items:
            type: string
          example:
            - APP
            - APV AND
            - MOBILE
            - WEB
        companies:
          type: array
          items:
            type: string
          example:
            - LP
            - SBB
        fulfillment_types:
          type: array
          items:
            type: string
          example:
            - Fulfillment_Type_Liverpool
            - Liverpool_CNC_PICK_PACK
        product_types:
          type: array
          items:
            type: string
          example:
            - Big Ticket
            - Soft Line
      required:
        - channels
        - companies
        - fulfillment_types
        - product_types

    StatItem:
      type: object
      properties:
        name:
          type: string
          example: WEB
        total:
          type: integer
          format: int64
          example: 28532
      required:
        - name
        - total

    StatsResponse:
      type: object
      properties:
        total_orders:
          type: integer
          format: int64
          example: 59848
        error_percentage:
          type: number
          format: float
          example: 12.02
        by_channel:
          type: array
          items:
            $ref: "#/components/schemas/StatItem"
        by_fulfillment_type:
          type: array
          items:
            $ref: "#/components/schemas/StatItem"
        by_product_type:
          type: array
          items:
            $ref: "#/components/schemas/StatItem"
      required:
        - total_orders
        - error_percentage
        - by_channel
        - by_fulfillment_type
        - by_product_type